app/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
jinja2
passlib[bcrypt]
python-jose
pydantic
psutil
aiofiles
python-multipart
requests
